+++ calculations-service/src/main/java/com/dzydowicz/calculations/application/service/CalculationsService.java
@@ -9,8 +9,8 @@ class CalculationsService implements CalculationsServicePort {
     private static final Integer VELOCITY = 17; //km/s
 
     @Override
-    public Double calculateKineticEnergy(Integer mass) {
-        return 0.5 * mass * Math.pow(VELOCITY, 2);
+    public Float calculateKineticEnergy(Float mass) {
+        return 0.5f * mass * (float) Math.pow(VELOCITY, 2);
     }
 
 }
